pat: increase current total key size when we use KEY_LARGE flag (#2368)
This commit is part of the work to implement
https://github.com/groonga/groonga/issues/2349.

Increase current total key size when we use `KEY_LARGE` flag by this
commit.
As a result, we also expand return value type to `uint64_t` from
`uint32_t`.
Because current total key size may over `uint32_t` when we use
`KEY_LARGE` flag.

---------

+++ include/groonga/groonga.h
@@ -1905,11 +1905,12 @@ grn_obj_flush_only_opened(grn_ctx *ctx, grn_obj *obj);
  * \param threshold The threshold value used to guide the defragmentation
  *                  process.
  *
- * \return The total number of segments on which defragmentation was performed.
+ * \return The total number of reduced bytes or segments on which
+ *         defragmentation was performed.
  *
  *         See `ctx->rc` for error details.
  */
-GRN_API int
+GRN_API uint64_t
 grn_obj_defrag(grn_ctx *ctx, grn_obj *obj, int threshold);
 
 /**
